pyproject.toml
[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''/(\n  # Directories to exclude\n  \.git\n  | \.venv\n  | venv\n  | build\n  | dist\n)/
'''

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
ensure_newline_before_comments = true
known_first_party = ["stocker"]
known_third_party = ["fastapi", "pydantic", "sqlalchemy", "pytest"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--strict-markers -v"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv"]
skips = ["B101"]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry]
name = "stocker-pro"
version = "1.0.0"
description = "STOCKER Pro - Advanced stock analysis and portfolio management"
authors = ["STOCKER Pro Team <team@stockerpro.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = "^0.22.0"
pydantic = "^1.10.7"
sqlalchemy = "^2.0.9"
alembic = "^1.10.3"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.0.1"
python-multipart = "^0.0.6"
email-validator = "^2.0.0"
psutil = "^5.9.5"
pandas = "^2.0.0"
numpy = "^1.24.3"
scipy = "^1.10.1"
matplotlib = "^3.7.1"
seaborn = "^0.12.2"
scikit-learn = "^1.2.2"
xgboost = "^1.7.5"
lightgbm = "^3.3.5"
tensorflow = "^2.12.0"
keras = "^2.12.0"
yfinance = "^0.2.18"
requests = "^2.28.2"
aiohttp = "^3.8.4"
prometheus-client = "^0.16.0"
python-json-logger = "^2.0.7"
pyjwt = "^2.6.0"
cryptography = "^40.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.0"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.2.0"
flake8 = "^6.0.0"
pre-commit = "^3.2.2"
mkdocs = "^1.4.3"
mkdocs-material = "^9.1.5"

[tool.poetry.scripts]
api = "stocker.interfaces.api.main:run_server"
